certsubmitter: Add SCT signature verification check
This PR also moves some structures, types and testdata to shared locations so
that it can be used for testing by multiple monologue packages.
